Copy DB script will now check for db's existence before replacing the file

+++ scripts/copy-db.sh
@@ -1,6 +1,6 @@
 # Help dialog
 help() {
-    echo Use: $0 [path]
+    echo Use: $0 [options] [path]
     echo Warning! Overrides existing files!
 }
 
@@ -28,10 +28,19 @@ innerpath="/opt/FileMaker/FileMaker Server/Data/Databases"
 cmd="docker exec -itu0 fms"
 
 # Copy and change permissions
+exists=$(docker exec fms [ -f "/$innerpath/$dbname" ] && echo true || echo false)
+if [ $exists = true ]; then
+    echo "This file already exists, would you like to replace it? [y/N]"
+    read replace
+    [ -z $replace ] || ([ $replace != "y" ] && [ $replace != "Y" ]) && exit 0
+    echo Replacing...
+fi
+
 docker cp "$dbname" "fms:$innerpath"
 $cmd chown fmserver:fmsadmin "$innerpath/$dbname"
 $cmd ls -l "$innerpath"
 
 # Possible plans:
 # - Prompt to change name
+# - Copy to Secure or other directories
 # - Allow -c flag to copy Devin commits?
